+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/BoundedChannelFullMode.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Threading.Tasks.Channels
+namespace System.IO.Channels
 {
     /// <summary>Specifies the behavior to use when writing to a bounded channel that is already full.</summary>
     public enum BoundedChannelFullMode
