+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/ChannelClosedException.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Threading.Tasks.Channels
+namespace System.IO.Channels
 {
     /// <summary>Exception thrown when a channel is used after it's been closed.</summary>
     public class ChannelClosedException : InvalidOperationException
