+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/ChannelOptimizations.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Threading.Tasks.Channels
+namespace System.IO.Channels
 {
     /// <summary>Provides settings for possible optimizations / constraints to use when creating a channel.</summary>
     public sealed class ChannelOptimizations
