+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/ChannelUtilities.cs
@@ -5,8 +5,10 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
 
-namespace System.Threading.Tasks.Channels
+namespace System.IO.Channels
 {
     /// <summary>Provides internal helper methods for implementing channels.</summary>
     internal static class ChannelUtilities
