+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/Channel_1.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Threading.Tasks.Channels
+namespace System.IO.Channels
 {
     /// <summary>Provides a base class for channels that support reading and writing elements of type <typeparamref name="T"/>.</summary>
     /// <typeparam name="T">Specifies the type of data readable and writable in the channel.</typeparam>
