+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/Interactor.cs
@@ -2,7 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Threading.Tasks.Channels
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace System.IO.Channels
 {
     internal abstract class Interactor<T> : TaskCompletionSource<T>
     {
