+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/SingleConsumerUnboundedChannel.cs
@@ -6,8 +6,10 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
 
-namespace System.Threading.Tasks.Channels
+namespace System.IO.Channels
 {
     /// <summary>
     /// Provides a buffered channel of unbounded capacity for use by any number
