+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/WritableChannel.cs
@@ -5,8 +5,10 @@
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
 
-namespace System.Threading.Tasks.Channels
+namespace System.IO.Channels
 {
     /// <summary>
     /// Provides a base class for channels that support writing elements.
