+++ src/System.Threading.Tasks.Channels/System/VoidResult.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Threading.Tasks
+namespace System
 {
     /// <summary>An empty struct, used to represent void in generic types.</summary>
     internal struct VoidResult { }
