+++ tests/System.Threading.Tasks.Channels.Tests/ChannelClosedExceptionTests.cs
@@ -3,7 +3,7 @@
 
 using Xunit;
 
-namespace System.Threading.Tasks.Channels.Tests
+namespace System.IO.Channels.Tests
 {
     public class ChannelClosedExceptionTests
     {
