+++ tests/System.Threading.Tasks.Channels.Tests/ChannelTests.cs
@@ -5,9 +5,11 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Xunit;
 
-namespace System.Threading.Tasks.Channels.Tests
+namespace System.IO.Channels.Tests
 {
     public class ChannelTests
     {
