+++ tests/System.Threading.Tasks.Channels.Tests/UnbufferedChannelTests.cs
@@ -3,9 +3,11 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Xunit;
 
-namespace System.Threading.Tasks.Channels.Tests
+namespace System.IO.Channels.Tests
 {
     public class UnbufferedChannelTests : ChannelTestBase
     {
