+++ tests/System.Threading.Tasks.Channels.Tests/ValueAwaiterTests.cs
@@ -4,9 +4,11 @@
 
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
 using Xunit;
 
-namespace System.Threading.Tasks.Channels.Tests
+namespace System.IO.Channels.Tests
 {
     public class ValueAwaiterTests
     {
